% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_utils.R
\name{make_title_col_where}
\alias{make_title_col_where}
\title{Make title vector of data column names}
\usage{
make_title_col_where(df, type)
}
\arguments{
\item{df}{dataset}

\item{type}{one of factor, double, integer, and character}
}
\value{
cols named vector with column names and title names
}
\description{
Make title vector of data column names
}
\examples{
make_title_col_where(mpg)
}
