% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot.R
\name{plot_server}
\alias{plot_server}
\title{plot Server}
\usage{
plot_server(id, var_inputs)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{var_inputs}{inputs from module_var_input.}
}
\value{
shiny server module
}
\description{
plot Server
}
\keyword{internal}
